+++ packages/uui-badge/lib/uui-badge.element.ts
@@ -2,8 +2,8 @@ import { defineElement } from '@umbraco-ui/uui-base/lib/registration';
 import { css, html, LitElement } from 'lit';
 import { property } from 'lit/decorators.js';
 import type {
-  InterfaceColor,
-  InterfaceLook,
+  UUIInterfaceColor,
+  UUIInterfaceLook,
 } from '@umbraco-ui/uui-base/lib/types';
 
 /**
@@ -21,7 +21,7 @@ export class UUIBadgeElement extends LitElement {
    * @default "default"
    */
   @property({ reflect: true })
-  color: InterfaceColor = 'default';
+  color: UUIInterfaceColor = 'default';
 
   /**
    * Changes the look of the button to one of the predefined, symbolic looks. For example - set this to positive if you want nice, green "confirm" button.
@@ -30,7 +30,7 @@ export class UUIBadgeElement extends LitElement {
    * @default "default"
    */
   @property({ reflect: true })
-  look: InterfaceLook = 'primary';
+  look: UUIInterfaceLook = 'primary';
 
   /**
    * Bring attention to this badge by applying a bounce animation.