+++ packages/uui-box/lib/uui-box.test.ts
@@ -1,7 +1,7 @@
 import { expect, fixture, html } from '@open-wc/testing';
 
 import { UUIBoxElement } from './uui-box.element';
-import { InterfaceHeadingValues } from '@umbraco-ui/uui-base/lib/types';
+import { UUIInterfaceHeadingValues } from '@umbraco-ui/uui-base/lib/types';
 
 describe('UUIBox', () => {
   let element: UUIBoxElement;
@@ -16,7 +16,7 @@ describe('UUIBox', () => {
   });
 
   it('passes the a11y audit', async () => {
-    for (const headlineVariant of InterfaceHeadingValues) {
+    for (const headlineVariant of UUIInterfaceHeadingValues) {
       element = await fixture(
         html` <uui-box
           headline="headline"