+++ packages/uui-tabs/lib/uui-tabs.story.ts
@@ -1,6 +1,10 @@
 import '.';
 import '@umbraco-ui/uui-icon/lib';
 import '@umbraco-ui/uui-icon-registry-essential/lib';
+import '@umbraco-ui/uui-button/lib';
+import '@umbraco-ui/uui-popover-container/lib';
+import '@umbraco-ui/uui-symbol-more/lib';
+import '@umbraco-ui/uui-input/lib';
 
 import { Story } from '@storybook/web-components';
 import { html } from 'lit';
@@ -64,7 +68,7 @@ export const WithBorders: Story = () => html`
     display: flex;
     --uui-tab-divider: var(--uui-color-divider-standalone);
     ">
-    <uui-tab-group style="display: flex;">
+    <uui-tab-group>
       <uui-tab label="content"> Content </uui-tab>
       <uui-tab label="Packages"> Packages </uui-tab>
       <uui-tab label="Media" active> Media </uui-tab>
@@ -81,9 +85,9 @@ export const Navbar: Story = () => html`
     style="
     display: flex;
     height: 60px;
-    font-size: 16px;
+    font-size: var(--uui-type-default-size);
     ">
-    <uui-tab-group style="display: flex;">
+    <uui-tab-group>
       <uui-tab label="content"> Content </uui-tab>
       <uui-tab label="Packages" active> Packages </uui-tab>
       <uui-tab label="Media"> Media </uui-tab>
@@ -100,7 +104,7 @@ export const UsingHref: Story = () => html`
     style="
     display: flex;
     height: 60px;
-    font-size: 16px;
+    font-size: var(--uui-type-default-size);
     ">
     <uui-tab-group>
       <uui-tab label="content" href="http://www.umbraco.com/#content">
@@ -132,18 +136,17 @@ export const WithIcons: Story = props => html`
       <uui-tab-group
         dropdown-direction="horizontal"
         style="
-        height: 70px;
-        font-size: 12px;
+        font-size: var(--uui-type-small-size);
         ${props.inlineStyles}">
-        <uui-tab>
+        <uui-tab label="content">
           <uui-icon slot="icon" name="document"></uui-icon>
           Content
         </uui-tab>
-        <uui-tab active>
+        <uui-tab active label="packages">
           <uui-icon slot="icon" name="settings"></uui-icon>
           Packages
         </uui-tab>
-        <uui-tab>
+        <uui-tab label="media">
           <uui-icon slot="icon" name="picture"></uui-icon>
           Media
         </uui-tab>
@@ -169,6 +172,229 @@ WithIcons.parameters = {
           Media
         </uui-tab>
       </uui-tab-group>`,
+    },
+  },
+};
+
+export const WithGap: Story = props => html`
+  <h3>Tabs with custom gap</h3>
+  <uui-icon-registry-essential>
+    <div style="display: flex;">
+      <uui-tab-group
+        dropdown-direction="horizontal"
+        style="
+        --uui-tab-group-gap: 180px;
+        margin: 0 auto;
+        font-size: var(--uui-type-small-size);
+        ${props.inlineStyles}">
+        <uui-tab label="content">
+          <uui-icon slot="icon" name="document"></uui-icon>
+          Content
+        </uui-tab>
+        <uui-tab active label="packages">
+          <uui-icon slot="icon" name="settings"></uui-icon>
+          Packages
+        </uui-tab>
+        <uui-tab label="media">
+          <uui-icon slot="icon" name="picture"></uui-icon>
+          Media
+        </uui-tab>
+      </uui-tab-group>
+    </div>
+  </uui-icon-registry-essential>
+`;
+WithGap.parameters = {
+  docs: {
+    source: {
+      code: `
+      <uui-tab-group>
+        <uui-tab>
+          <uui-icon slot="icon" name="document"></uui-icon>
+          Content
+        </uui-tab>
+        <uui-tab active>
+          <uui-icon slot="icon" name="settings"></uui-icon>
+          Packages
+        </uui-tab>
+        <uui-tab>
+          <uui-icon slot="icon" name="picture"></uui-icon>
+          Media
+        </uui-tab>
+      </uui-tab-group>`,
+    },
+  },
+};
+
+export const Async: Story = props => {
+  // async insert tabs after 1 second
+  setTimeout(() => {
+    const tabs = document.querySelector('uui-tab-group');
+
+    if (!tabs) return;
+
+    const tab = document.createElement('uui-tab');
+    tab.label = 'Async';
+    tab.innerHTML = 'Async';
+    tabs.appendChild(tab);
+
+    setTimeout(() => {
+      tab.innerHTML = 'Async more text';
+    }, 1000);
+  }, 1000);
+
+  return html`
+    <uui-icon-registry-essential>
+      <div style="display: flex">
+        <uui-tab-group
+          dropdown-direction="horizontal"
+          style="
+          margin: auto;
+        --uui-tab-group-gap: 25px;
+        font-size: var(--uui-type-small-size);
+        ${props.inlineStyles}">
+          <uui-tab label="content">
+            <uui-icon slot="icon" name="document"></uui-icon>
+            Content
+          </uui-tab>
+          <uui-tab active label="packages">
+            <uui-icon slot="icon" name="settings"></uui-icon>
+            Packages
+          </uui-tab>
+          <uui-tab label="media">
+            <uui-icon slot="icon" name="picture"></uui-icon>
+            Media
+          </uui-tab>
+        </uui-tab-group>
+      </div>
+    </uui-icon-registry-essential>
+  `;
+};
+Async.parameters = {
+  docs: {
+    source: {
+      code: `
+      <uui-tab-group>
+        <uui-tab>
+          <uui-icon slot="icon" name="document"></uui-icon>
+          Content
+        </uui-tab>
+        <uui-tab active>
+          <uui-icon slot="icon" name="settings"></uui-icon>
+          Packages
+        </uui-tab>
+        <uui-tab>
+          <uui-icon slot="icon" name="picture"></uui-icon>
+          Media
+        </uui-tab>
+      </uui-tab-group>`,
+    },
+  },
+};
+
+export const FlexLayout: Story = props => html`
+  <uui-icon-registry-essential>
+    <div style="display: flex;">
+      <uui-input style="min-width: 200px; flex-grow: 1"></uui-input>
+      <uui-tab-group
+        dropdown-direction="horizontal"
+        style="
+        ${props.inlineStyles}">
+        <uui-tab label="content">Content</uui-tab>
+        <uui-tab active label="packages">Packages</uui-tab>
+        <uui-tab label="media">Media</uui-tab>
+        <uui-tab label="settings">Settings</uui-tab>
+        <uui-tab label="translations">Translations</uui-tab>
+      </uui-tab-group>
+    </div>
+  </uui-icon-registry-essential>
+`;
+FlexLayout.parameters = {
+  docs: {
+    source: {
+      code: `
+      <div style="display: flex">
+        <uui-tab-group style="margin: auto">
+          <uui-tab label="content">Content</uui-tab>
+          <uui-tab active label="packages">Packages</uui-tab>
+          <uui-tab label="media">Media</uui-tab>
+          <uui-tab label="settings">Settings</uui-tab>
+          <uui-tab label="translations">Translations</uui-tab>
+          </uui-tab-group>
+      </div>
+      `,
+    },
+  },
+};
+
+export const CenterAlign: Story = props => html`
+  <h3>Tabs implemented into Flex-box scenario</h3>
+  <p>Here the tab group is center aligned in a flex-box container.</p>
+  <uui-icon-registry-essential>
+    <div style="display: flex; justify-content: center">
+      <uui-tab-group
+        dropdown-direction="horizontal"
+        style="--uui-tab-group-gap: 25px;
+        ${props.inlineStyles}">
+        <uui-tab label="content">Content</uui-tab>
+        <uui-tab active label="packages">Packages</uui-tab>
+        <uui-tab label="media">Media</uui-tab>
+        <uui-tab label="settings">Settings</uui-tab>
+        <uui-tab label="translations">Translations</uui-tab>
+      </uui-tab-group>
+    </div>
+  </uui-icon-registry-essential>
+`;
+CenterAlign.parameters = {
+  docs: {
+    source: {
+      code: `
+      <div style="display: flex">
+        <uui-tab-group style="margin: auto">
+          <uui-tab label="content">Content</uui-tab>
+          <uui-tab active label="packages">Packages</uui-tab>
+          <uui-tab label="media">Media</uui-tab>
+          <uui-tab label="settings">Settings</uui-tab>
+          <uui-tab label="translations">Translations</uui-tab>
+          </uui-tab-group>
+      </div>
+      `,
+    },
+  },
+};
+
+export const RightAlign: Story = props => html`
+  <h3>Tabs implemented into Flex-box scenario</h3>
+  <p>Here the tab group is right aligned in a flex-box container.</p>
+  <uui-icon-registry-essential>
+    <div style="display: flex; justify-content: right">
+      <uui-tab-group
+        dropdown-direction="horizontal"
+        style="
+        --uui-tab-group-gap: 25px;
+        ${props.inlineStyles}">
+        <uui-tab label="content">Content</uui-tab>
+        <uui-tab active label="packages">Packages</uui-tab>
+        <uui-tab label="media">Media</uui-tab>
+        <uui-tab label="settings">Settings</uui-tab>
+        <uui-tab label="translations">Translations</uui-tab>
+      </uui-tab-group>
+    </div>
+  </uui-icon-registry-essential>
+`;
+RightAlign.parameters = {
+  docs: {
+    source: {
+      code: `
+      <div style="display: flex">
+        <uui-tab-group style="margin: auto">
+          <uui-tab label="content">Content</uui-tab>
+          <uui-tab active label="packages">Packages</uui-tab>
+          <uui-tab label="media">Media</uui-tab>
+          <uui-tab label="settings">Settings</uui-tab>
+          <uui-tab label="translations">Translations</uui-tab>
+          </uui-tab-group>
+      </div>
+      `,
     },
   },
 };