+++ CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0...v1.6.1) (2024-02-06)
+
+### Bug Fixes
+
+- **uui-tab-group:** layout is causing issues due to missing `display: flex` ([#730](https://github.com/umbraco/Umbraco.UI/issues/730)) ([5818804](https://github.com/umbraco/Umbraco.UI/commit/58188040d30f872a6ee1c68f92879e3c0e3ac1b1))
+
+# [1.6.0](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0-rc.4...v1.6.0) (2024-02-02)
+
+### Bug Fixes
+
+- :bug: swap add icon svg storing sor a correct one ([#727](https://github.com/umbraco/Umbraco.UI/issues/727)) ([9b64e2d](https://github.com/umbraco/Umbraco.UI/commit/9b64e2de1758067c029e28e73f31207485a21fa5))
+- uui-tab-group should have display set to block ([#722](https://github.com/umbraco/Umbraco.UI/issues/722)) ([ac5c2b1](https://github.com/umbraco/Umbraco.UI/commit/ac5c2b18b183fc8dcf9a9b2d641782db3f9215ad))
+- **uui-tab-group:** async tabs are not calculated correctly and have memory leaks ([#726](https://github.com/umbraco/Umbraco.UI/issues/726)) ([03c5cdc](https://github.com/umbraco/Umbraco.UI/commit/03c5cdc794b316ce290e7db27860ac8043681c7c))
+
+# [1.6.0-rc.4](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0-rc.3...v1.6.0-rc.4) (2024-01-23)
+
+### Bug Fixes
+
+- direct imports are not allowed ([3b4b4dd](https://github.com/umbraco/Umbraco.UI/commit/3b4b4dd2ade9b531cde1d35e2808b96b485eb3c7))
+- use css var for text ([#717](https://github.com/umbraco/Umbraco.UI/issues/717)) ([278a6dd](https://github.com/umbraco/Umbraco.UI/commit/278a6ddac8556ee94a254aa2278654d4a264fde2))
+- use css var for text ([#718](https://github.com/umbraco/Umbraco.UI/issues/718)) ([f5ccaa4](https://github.com/umbraco/Umbraco.UI/commit/f5ccaa42c4e5b07bf89226b5de835fa529161840))
+- uui-popover container scroll event ([#713](https://github.com/umbraco/Umbraco.UI/issues/713)) ([656f181](https://github.com/umbraco/Umbraco.UI/commit/656f1811a1de862c7a7f6ca0221bf14e3532d53c))
+- uui-tab-group support for gap ([#712](https://github.com/umbraco/Umbraco.UI/issues/712)) ([2f07793](https://github.com/umbraco/Umbraco.UI/commit/2f07793d7a9293b35dcc221006fee7be8c5df79c))
+
+### Features
+
+- Inline-button-create support for HREF ([#720](https://github.com/umbraco/Umbraco.UI/issues/720)) ([7d8affd](https://github.com/umbraco/Umbraco.UI/commit/7d8affdcd7d68e3a68a06a1591084b61ac5fe068))
+- Prefix interface types with UUI ([#716](https://github.com/umbraco/Umbraco.UI/issues/716)) ([7cd20db](https://github.com/umbraco/Umbraco.UI/commit/7cd20dbcbef51fbee308798a90e3bb4c69dbf503))
+- uui-ref-node should have href and target properties ([#715](https://github.com/umbraco/Umbraco.UI/issues/715)) ([12feafb](https://github.com/umbraco/Umbraco.UI/commit/12feafbbef3f736623503d452332eef7a5dea073))
+
 # [1.6.0-rc.3](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0-rc.2...v1.6.0-rc.3) (2023-12-15)
 
 ### Bug Fixes