+++ packages/uui-box/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0-rc.4...v1.6.0) (2024-02-02)
+
+**Note:** Version bump only for package @umbraco-ui/uui-box
+
+# [1.6.0-rc.4](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0-rc.3...v1.6.0-rc.4) (2024-01-23)
+
+### Features
+
+- Prefix interface types with UUI ([#716](https://github.com/umbraco/Umbraco.UI/issues/716)) ([7cd20db](https://github.com/umbraco/Umbraco.UI/commit/7cd20dbcbef51fbee308798a90e3bb4c69dbf503))
+
 # [1.6.0-rc.3](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0-rc.2...v1.6.0-rc.3) (2023-12-15)
 
 **Note:** Version bump only for package @umbraco-ui/uui-box