+++ packages/uui-button-inline-create/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0-rc.4...v1.6.0) (2024-02-02)
+
+**Note:** Version bump only for package @umbraco-ui/uui-button-inline-create
+
+# [1.6.0-rc.4](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0-rc.3...v1.6.0-rc.4) (2024-01-23)
+
+### Features
+
+- Inline-button-create support for HREF ([#720](https://github.com/umbraco/Umbraco.UI/issues/720)) ([7d8affd](https://github.com/umbraco/Umbraco.UI/commit/7d8affdcd7d68e3a68a06a1591084b61ac5fe068))
+
 # [1.6.0-rc.3](https://github.com/umbraco/Umbraco.UI/compare/v1.6.0-rc.2...v1.6.0-rc.3) (2023-12-15)
 
 **Note:** Version bump only for package @umbraco-ui/uui-button-inline-create